requirements.txt
# Core
Django==4.2.7
Pillow>=10.3.0
django-environ==0.11.2
whitenoise==6.6.0
openpyxl==3.1.2
requests==2.31.0

# REST
djangorestframework==3.15.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2

# Admin
django-jazzmin

# WebSocket
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0

# Celery + Beat
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# PDF
reportlab>=4.0.0

# Telegram Bot
python-telegram-bot>=20.0  # optional, hozir requests ishlatilmoqda

# Production Server & Packages
psycopg2-binary
matplotlib
opencv-python
pyzbar
gunicorn==21.2.0
uvicorn==0.24.0